+++ app/code/local/Stuntcoders/FbPixel/Model/Observer.php
@@ -1,0 +1,9 @@
+<?php
+
+class Stuntcoders_FbPixel_Model_Observer
+{
+    public function onAddToWishlist()
+    {
+        Mage::getSingleton('customer/session')->setIsProductAddedToWishlist(true);
+    }
+}